refactor some logic into CSV Form from App
the whole logic of parsing, validation and submitting transactions can
be handled by the CSVForm and thus removed from the App
