+++ src/App.tsx
@@ -1,78 +1,17 @@
-import { SafeAppProvider } from "@gnosis.pm/safe-apps-provider";
-import { useSafeAppsSDK } from "@gnosis.pm/safe-apps-react-sdk";
 import { Loader, Text } from "@gnosis.pm/safe-react-components";
-import { useWhatChanged, setUseWhatChange } from "@simbathesailor/use-what-changed";
-import { ethers } from "ethers";
-import debounce from "lodash.debounce";
-import React, { useCallback, useState, useContext, useMemo } from "react";
+import { setUseWhatChange } from "@simbathesailor/use-what-changed";
+import React from "react";
 import styled from "styled-components";
 
 import { CSVForm } from "./components/CSVForm";
 import { Header } from "./components/Header";
-import { MessageContext } from "./contexts/MessageContextProvider";
-import { useTokenInfoProvider, useTokenList } from "./hooks/token";
-import { parseCSV, Payment } from "./parser";
-import { buildTransfers } from "./transfers";
-import { checkAllBalances, transfersToSummary } from "./utils";
+import { useTokenList } from "./hooks/token";
 
 setUseWhatChange(process.env.NODE_ENV === "development");
 
 const App: React.FC = () => {
-  const { safe, sdk } = useSafeAppsSDK();
+  const { isLoading } = useTokenList();
 
-  const { tokenList, isLoading } = useTokenList();
-  const tokenInfoProvider = useTokenInfoProvider();
-  const [submitting, setSubmitting] = useState(false);
-  const [parsing, setParsing] = useState(false);
-  const [transferContent, setTransferContent] = useState<Payment[]>([]);
-  const [csvText, setCsvText] = useState<string>("token_address,receiver,amount");
-  const { setCodeWarnings, setMessages } = useContext(MessageContext);
-
-  const web3Provider = useMemo(() => new ethers.providers.Web3Provider(new SafeAppProvider(safe, sdk)), [safe, sdk]);
-
-  const parseAndValidateCSV = useMemo(
-    () =>
-      debounce((csvText: string) => {
-        setParsing(true);
-        const parsePromise = parseCSV(csvText, tokenInfoProvider);
-        parsePromise
-          .then(([transfers, warnings]) => {
-            console.log("CSV parsed!");
-            const summary = transfersToSummary(transfers);
-            checkAllBalances(summary, web3Provider, safe).then((insufficientBalances) =>
-              setMessages(
-                insufficientBalances.map((insufficientBalanceInfo) => ({
-                  message: `Insufficient Balance: ${insufficientBalanceInfo.transferAmount} of ${insufficientBalanceInfo.token}`,
-                  severity: "warning",
-                })),
-              ),
-            );
-            setTransferContent(transfers);
-            setCodeWarnings(warnings);
-            setParsing(false);
-          })
-          .catch((reason: any) => setMessages([{ severity: "error", message: reason.message }]));
-      }, 1000),
-    [safe, setCodeWarnings, setMessages, tokenInfoProvider, web3Provider],
-  );
-  const onChangeTextHandler = (csvText: string) => {
-    setCsvText(csvText);
-    parseAndValidateCSV(csvText);
-  };
-
-  const submitTx = useCallback(async () => {
-    setSubmitting(true);
-    try {
-      const txs = buildTransfers(transferContent);
-      console.log(`Encoded ${txs.length} ERC20 transfers.`);
-      const sendTxResponse = await sdk.txs.send({ txs });
-      const safeTx = await sdk.txs.getBySafeTxHash(sendTxResponse.safeTxHash);
-      console.log({ safeTx });
-    } catch (e) {
-      console.error(e);
-    }
-    setSubmitting(false);
-  }, [transferContent, sdk.txs]);
   return (
     <Container>
       <Header />
@@ -82,16 +21,7 @@ const App: React.FC = () => {
           <Text size={"lg"}>Loading Tokenlist...</Text>
         </>
       ) : (
-        <CSVForm
-          csvText={csvText}
-          onAbortSubmit={() => setSubmitting(false)}
-          submitting={submitting}
-          transferContent={transferContent}
-          onSubmit={submitTx}
-          onChange={onChangeTextHandler}
-          tokenList={tokenList}
-          submitEnabled={!parsing}
-        />
+        <CSVForm />
       )}
     </Container>
   );
